requirements.txt
# Core application dependencies for Railway deployment
# Migrated to PostgreSQL + pgvector

# Unstructured - for document parsing (with PDF support)
unstructured[pdf]==0.18.15

# Web framework and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# LangChain ecosystem - compatible versions verified
langchain==0.3.7
langchain-community==0.3.7
langchain-openai==0.2.9
langchain-core==0.3.17
langchain-text-splitters==0.3.2

# PostgreSQL and pgvector support (using langchain-community's PGVector)
psycopg2-binary==2.9.9
sqlalchemy==2.0.32
pgvector==0.3.2
psycopg[binary]==3.2.3  # Required by langchain-community PGVector

# Document loaders
pypdf==4.3.1
docx2txt==0.8

# Enhanced PDF processing with OCR
PyMuPDF==1.24.10
pdf2image==1.17.0
pytesseract==0.3.13
Pillow==10.4.0

# OpenAI API - compatible with langchain-openai 0.2.9
openai==1.54.4

# Google Drive integration
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.149.0

# Utilities
python-dotenv==1.0.1
tiktoken==0.8.0
pyyaml==6.0.1

# Compatibility pins
pydantic==2.9.2
pydantic-core==2.23.4

